Avoid casting by using s8 throughout (#7)

+++ sound/pci/hdsp/hdspe/hdspe_mixer.c
@@ -111,7 +111,7 @@ void hdspe_mixer_update_channel_map(struct hdspe* hdspe)
 		used[i] = false;
 	}
 	for (i = 0; i < HDSPE_MIXER_CHANNELS; i ++) {
-		const int c = hdspe->channel_map_out[i];
+		const s8 c = hdspe->channel_map_out[i];
 		if (c >= 0 && c < HDSPE_MAX_CHANNELS)
 			used[c] = true;
 	}
@@ -133,7 +133,7 @@ void hdspe_mixer_update_channel_map(struct hdspe* hdspe)
 		used[i] = false;
 	}
 	for (i = 0; i < HDSPE_MIXER_CHANNELS; i ++) {
-		const int c = hdspe->channel_map_in[i];
+		const s8 c = hdspe->channel_map_in[i];
 		if (c >= 0 && c < HDSPE_MAX_CHANNELS)
 			used[c] = true;
 	}
